docs: update README with new modular layout
