+++ android/__init__.py
@@ -1,0 +1,5 @@
+"""Android (Kivy) GUI placeholder package."""
+
+from .main import demo_scan, main
+
+__all__ = ["demo_scan", "main"]
