+++ android/kv/main.kv
@@ -1,0 +1,9 @@
+#:kivy 2.0.0
+
+# Placeholder KV layout for future Android UI.
+
+<RootWidget>:
+    BoxLayout:
+        orientation: 'vertical'
+        Label:
+            text: 'Android UI placeholder'
