+++ core/__init__.py
@@ -1,0 +1,25 @@
+"""Shared logic for the Tasmota Bulk Tool."""
+
+from .utils import build_ip_list, format_log_line, timestamp_now
+from .parser import (
+    DeviceResult,
+    safe_extract_json,
+    parse_status_payload,
+    normalize_command_library,
+)
+from .network import http_get, send_command, collect_device_info
+from .ota import perform_ota_upgrade
+
+__all__ = [
+    "build_ip_list",
+    "format_log_line",
+    "timestamp_now",
+    "DeviceResult",
+    "safe_extract_json",
+    "parse_status_payload",
+    "normalize_command_library",
+    "http_get",
+    "send_command",
+    "collect_device_info",
+    "perform_ota_upgrade",
+]
