+++ desktop/__init__.py
@@ -1,0 +1,1 @@
+"""Desktop (PySide6) GUI package."""
